requirements.txt
google-cloud-bigquery==3.13.0
google-cloud-storage==2.14.0
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
dbt-bigquery==1.7.0
scikit-learn==1.4.0
xgboost==2.0.3
google-cloud-aiplatform==1.38.0
streamlit==1.31.0
flask==3.1.0
db-dtypes
setuptools>=68.0.0
